+++ src/transformers/__init__.py
@@ -19,6 +19,18 @@ else:
 
 import logging
 
+# Benchmarking
+from .benchmark_utils import (
+    Frame,
+    Memory,
+    MemoryState,
+    MemorySummary,
+    MemoryTrace,
+    UsedMemoryState,
+    start_memory_tracing,
+    stop_memory_tracing,
+    bytes_to_human_readable,
+)
 from .configuration_albert import ALBERT_PRETRAINED_CONFIG_ARCHIVE_MAP, AlbertConfig
 from .configuration_auto import ALL_PRETRAINED_CONFIG_ARCHIVE_MAP, AutoConfig
 from .configuration_bart import BartConfig
@@ -71,12 +83,9 @@ from .file_utils import (
     WEIGHTS_NAME,
     add_end_docstrings,
     add_start_docstrings,
-    bytes_to_human_readable,
     cached_path,
     is_tf_available,
     is_torch_available,
-    start_memory_tracing,
-    stop_memory_tracing,
 )
 
 # Model Cards
